[Chore] 후기 작성 완료한 책일 시 인생책 삭제하면 IsFavorite필드값만 false로 변경

+++ src/main/java/com/example/bookiibookii/domain/user/service/BookshelfService.java
@@ -164,6 +164,13 @@ public class BookshelfService {
             throw new UserException(UserErrorCode.USER_BOOK_NOT_FOUND);
         }
 
-        userBookRepository.delete(userBook);
+        boolean shouldKeep = userBook.getDisplayOrder() != null
+                || groupBookRepository.existsByUser_IdAndBook_IdAndRatingIsNotNull(userId, userBook.getBook().getId());
+
+        if (shouldKeep) {
+            userBook.updateIsFavorite(false);
+        } else {
+            userBookRepository.delete(userBook);
+        }
     }
 }
